backend/requirements.txt
Django==4.1.13
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt
gunicorn==20.1.0